market_making/requirements.txt

redis
python-socketio 
asyncio 
aiohttp